change name database to domain
