+++ freeapply/domain/__init__.py
@@ -1,2 +1,2 @@
 
-from .database import *
+from .domain import *
